Implement Poseidon tweak and message hash (#4)
- Added num-bigint dependency
- Chain lengths and indices are now u16 to have a smaller tweak
- MESSAGE_LENGTH changed to 32 for smaller hash
- Tweak separator values added
- Poseidon Message Hash now has TWEAK_LEN_FE and MSG_LEN_FE parameters
- Message encoded for hashing in Poseidon message hash
- Epoch/tweak encoded for hashing in Poseidon message hash
- Poseidon message hash and weakable hash implementation
- Introduce the concept of consistency check function for testing purposes only

+++ src/inc_encoding.rs
@@ -41,7 +41,7 @@ pub trait IncomparableEncoding {
         message: &[u8; MESSAGE_LENGTH],
         randomness: &Self::Randomness,
         epoch: u32,
-    ) -> Result<Vec<u32>, EncodingError>;
+    ) -> Result<Vec<u16>, EncodingError>;
 }
 
 pub mod basic_winternitz;
